Specify image dimensions in CSS file
Previously, the dimensions of the images of the prepared
dishes were only defined in the HTML files. Now, the same
dimensions have also been defined in the CSS file, to ensure
a high level of specificity, and that physical space is allocated
on each page for the image, before the image loads.